WIP: split trip metadata tests for LOC (checkpoint)
